Remove unnecessary `PushChannelNative` class
